Fix web beta gate review findings
