+++ apps/packages/product-ui/src/auth/AuthStartPanel.tsx
@@ -1,7 +1,5 @@
-import { useState, type ReactNode } from "react";
-import { ChevronDown } from "@proliferate/ui/icons";
+import { type ReactNode } from "react";
 import { AuthProviderButton } from "@proliferate/ui/primitives/AuthProviderButton";
-import { Button } from "@proliferate/ui/primitives/Button";
 import { AuthLayout } from "./AuthLayout";
 
 export interface AuthProviderActionView {
@@ -21,9 +19,6 @@ interface AuthStartPanelProps {
   footer: ReactNode;
   providers: AuthProviderActionView[];
   credentialForm?: ReactNode;
-  secondaryActions?: AuthProviderActionView[];
-  secondaryLabel?: ReactNode;
-  secondaryContent?: ReactNode;
   note?: ReactNode;
   error?: ReactNode;
   devAccess?: ReactNode;
@@ -36,20 +31,14 @@ export function AuthStartPanel({
   footer,
   providers,
   credentialForm,
-  secondaryActions,
-  secondaryLabel = "More",
-  secondaryContent,
   note,
   error,
   devAccess,
 }: AuthStartPanelProps) {
-  const hasSecondaryOptions =
-    Boolean(secondaryActions?.length) || secondaryContent !== undefined;
-
   return (
     <AuthLayout mark={mark} title={title} subtitle={subtitle} footer={footer}>
-      {!hasSecondaryOptions ? credentialForm : null}
-      {!hasSecondaryOptions && credentialForm ? <AuthDivider /> : null}
+      {credentialForm}
+      {credentialForm ? <AuthDivider /> : null}
       {providers.map((provider) => (
         <AuthProviderButton
           key={provider.id}
@@ -62,13 +51,6 @@ export function AuthStartPanel({
           {provider.label}
         </AuthProviderButton>
       ))}
-      {hasSecondaryOptions ? (
-        <AuthSecondaryOptions
-          label={secondaryLabel}
-          actions={secondaryActions ?? []}
-          content={secondaryContent}
-        />
-      ) : null}
       {note ? <p className="mt-2 text-xs leading-5 text-muted-foreground">{note}</p> : null}
       {error ? (
         <div
@@ -80,60 +62,6 @@ export function AuthStartPanel({
       ) : null}
       {devAccess}
     </AuthLayout>
-  );
-}
-
-function AuthSecondaryOptions({
-  label,
-  actions,
-  content,
-}: {
-  label: ReactNode;
-  actions: AuthProviderActionView[];
-  content?: ReactNode;
-}) {
-  const [expanded, setExpanded] = useState(false);
-
-  return (
-    <div className="space-y-3">
-      <Button
-        type="button"
-        variant="secondary"
-        size="md"
-        aria-expanded={expanded}
-        className="h-11 w-full justify-center gap-2.5"
-        onClick={() => setExpanded((current) => !current)}
-      >
-        <span>{label}</span>
-        <ChevronDown
-          className={`size-3.5 shrink-0 text-muted-foreground transition-transform ${
-            expanded ? "rotate-180" : ""
-          }`}
-        />
-      </Button>
-      {expanded ? (
-        <div className="space-y-3">
-          {actions.map((action) => (
-            <AuthProviderButton
-              key={action.id}
-              icon={action.icon}
-              loading={action.loading}
-              disabled={action.disabled}
-              variant={action.primary ? "primary" : "secondary"}
-              onClick={action.onClick}
-            >
-              {action.label}
-            </AuthProviderButton>
-          ))}
-          {content ? (
-            <>
-              {actions.length > 0 ? <AuthDivider /> : null}
-              {content}
-            </>
-          ) : null}
-        </div>
-      ) : null}
-    </div>
   );
 }
