+++ server/proliferate/auth/identity/service.py
@@ -201,7 +201,8 @@ async def complete_oauth_provider_callback(
         ),
     )
     if callback_surface == "web" and challenge.purpose == "login":
-        ensure_web_beta_email_allowed(verified.email)
+        beta_email = await _beta_email_for_provider_login(db, verified=verified)
+        ensure_web_beta_email_allowed(beta_email)
     desktop_github_account_or_email_exists = True
     if callback_surface == "desktop" and provider == "github":
         desktop_github_account_or_email_exists = await _desktop_github_account_or_email_exists(
@@ -340,7 +341,8 @@ async def complete_apple_web_callback(
         display_name_hint=display_name,
     )
     if challenge.purpose == "login":
-        ensure_web_beta_email_allowed(verified.email)
+        beta_email = await _beta_email_for_provider_login(db, verified=verified)
+        ensure_web_beta_email_allowed(beta_email)
     user = await resolve_provider_user(db, verified=verified, challenge=challenge)
     auth_code = await create_auth_code(
         db,
@@ -351,6 +353,28 @@ async def complete_apple_web_callback(
         redirect_uri=challenge.redirect_uri,
     )
     return append_query(challenge.redirect_uri, code=auth_code.code, state=challenge.client_state)
+
+
+async def _beta_email_for_provider_login(
+    db: AsyncSession,
+    *,
+    verified: VerifiedProviderIdentity,
+) -> str | None:
+    existing_identity = await get_identity_by_provider_subject(
+        db,
+        provider=verified.provider,
+        provider_subject=verified.provider_subject,
+    )
+    if existing_identity is not None:
+        user = await get_user_by_id(db, existing_identity.user_id)
+        return user.email if user is not None else None
+
+    if verified.provider == "github" and verified.email:
+        existing_email_user = await get_user_by_email(db, verified.email)
+        if existing_email_user is not None:
+            return existing_email_user.email
+
+    return verified.email
 
 
 def _nonce_unavailable_marker(challenge: AuthChallengeSnapshot) -> str:
